drivers: npmx: change backend handling
Backend is passed as pointer to write and read functions.
It is more generic and does not require separted files
like backend_twim.c or backedn_spim.c.
